refactory: Separa responsabilidades dos services
