+++ src/service/index.ts
@@ -1,3 +1,4 @@
 export * from './Tree';
 export * from './Notification';
+export * from './Invitation';
 export * from './DynamicLink';
